% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{arguments}
\alias{arguments}
\title{Definition of the function arguments}
\arguments{
\item{lat1}{The latitude of point 1.}

\item{long1}{The longitude of point 1.}

\item{lat2}{The latitude of point 2.}

\item{long2}{The longitude of point 2.}
}
\description{
Here is the list of most function arguments used in the package.
The arguments not contained in this list are those for which the exact meaning depends on the context.
}
