% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{best_fit}
\alias{best_fit}
\title{The fit of the full (best) model}
\format{
a \code{HLfit} object produced by spaMM
}
\description{
This object contains the model fitted with the function \code{\link[=fit]{fit()}} as shown in \code{?mallaRd}.
}
\examples{
best_fit

}
\references{
this study
}
\seealso{
\code{\link[=fit]{fit()}}
}
\keyword{model}
