% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{compare_fits}
\alias{compare_fits}
\title{Compare the best fit to other fits using parametric bootstrap}
\usage{
compare_fits(fits, data, ncpus = 2, boot.repl = 1000, seed = 123)
}
\arguments{
\item{fits}{the table of fitted models produced by \code{fit()}}

\item{data}{the dataframe containing the data used by the function}

\item{ncpus}{the number of CPU cores to use for parallel processing (default = 2)}

\item{boot.repl}{the number of bootstrap replicates}

\item{seed}{the seed to be used for bootstrapping}
}
\value{
a dataframe with the formulas and all the statistics of goodness of fit
}
\description{
This function compute all the information required for model comparison.
}
\examples{
\dontrun{
  threeformulas <- prepare_formulas()[1:3,]
  fits <- fit(formulas = threeformulas, data = data_model)
  comparison <- compare_fits(fits, data = data_model, ncpus = 20, boot.repl = 100)
  comparison
}

}
