% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{compute_LRT}
\alias{compute_LRT}
\title{Compute Likelihood Ratio Test using parametric bootstrap}
\usage{
compute_LRT(fit, fit_null, ncpus = 2, boot.repl = 1000, seed = 123)
}
\arguments{
\item{fit}{a fitted model}

\item{fit_null}{a fitted model more simple than the other fit it is compared to}

\item{ncpus}{the number of CPU cores to use for parallel processing (default = 2)}

\item{boot.repl}{the number of bootstrap replicates}

\item{seed}{the seed to be used for bootstrapping}
}
\value{
the output of \link[spaMM:anova.HLfit]{spaMM::LRT}
}
\description{
Compute Likelihood Ratio Test using parametric bootstrap
}
