% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{compute_LRT2}
\alias{compute_LRT2}
\title{Compute Likelihood Ratio Test using parametric bootstrap for non nested models}
\usage{
compute_LRT2(best_fit, fit_alt, ncpus = 2, boot.repl = 1000, seed = 123)
}
\arguments{
\item{best_fit}{the fit of the model with the best predictive power}

\item{fit_alt}{an alternative fitted model to be compared to \code{best_fit}}

\item{ncpus}{the number of CPU cores to use for parallel processing (default = 2)}

\item{boot.repl}{the number of bootstrap replicates}

\item{seed}{the seed to be used for bootstrapping}
}
\value{
the output of \link[spaMM:anova.HLfit]{spaMM::LRT}
}
\description{
Compute Likelihood Ratio Test using parametric bootstrap for non nested models
}
