% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_all}
\alias{data_all}
\title{The cleaned dataset of all breeding events}
\format{
a \code{tibble}
}
\description{
This dataset contains the raw data after reformatting the variables but without any filtering. It
derives from \code{data_raw} as shown in \code{?mallaRd}.
}
\examples{
data_all

}
\references{
this study
}
\seealso{
\link{data_raw}, \link{data_model}
}
\keyword{datasets}
