% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_all_known}
\alias{data_all_known}
\title{The cleaned dataset of all breeding events for identified birds}
\format{
a \code{tibble}
}
\description{
This dataset contains the raw data after reformatting the variables and after discarding the unknown birds.
It derives from \code{data_all} as shown in \code{?mallaRd}.
}
\examples{
data_all_known

}
\references{
this study
}
\seealso{
\link{data_raw}, \link{data_model}
}
\keyword{datasets}
