% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_model}
\alias{data_model}
\title{The complete dataset for fitting the models}
\format{
a \code{tibble}
}
\description{
This dataset contains the data used to fit all the models.
}
\examples{
data_model

}
\references{
this study
}
\keyword{datasets}
