% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_raw}
\alias{data_raw}
\title{The raw dataset of all breeding events}
\format{
a \code{data.frame}
}
\description{
This dataset contains the raw data, unprocessed.
}
\examples{
data_raw

}
\references{
this study
}
\seealso{
\link{data_all}, \link{data_model}
}
\keyword{datasets}
