% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{figure1}
\alias{figure1}
\title{Build figure 1}
\usage{
figure1(data)
}
\arguments{
\item{data}{the dataframe containing the data used by the function}
}
\description{
Build figure 1
}
\examples{
figure1(data_all)

}
