% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{figure2}
\alias{figure2}
\title{Build figure 2}
\usage{
figure2(data)
}
\arguments{
\item{data}{the dataframe containing the data used by the function}
}
\description{
Build figure 2
}
\examples{
figure2(data_all)

}
