% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{figure3}
\alias{figure3}
\title{Build figure 3}
\usage{
figure3(data, data_model)
}
\arguments{
\item{data}{the dataframe containing the data used by the function}

\item{data_model}{the dataframe containing the data used for model fitting}
}
\description{
Build figure 3
}
\examples{
figure3(data_all, data_model)

}
