% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{figure4}
\alias{figure4}
\title{Build figure 4}
\usage{
figure4(data_model, fit, ...)
}
\arguments{
\item{data_model}{the dataframe containing the data used for model fitting}

\item{fit}{a fitted model}

\item{...}{addition arguments for internal functions}
}
\description{
Build figure 4
}
\examples{
figure4(data_model, best_fit)

}
