% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{figure5}
\alias{figure5}
\title{Build figure 5}
\usage{
figure5(best_fit)
}
\arguments{
\item{best_fit}{the fit of the model with the best predictive power}
}
\description{
Build figure 5
}
\examples{
figure5(best_fit)

}
