% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{figureS1}
\alias{figureS1}
\title{Build figure S1}
\usage{
figureS1(data, ymax = c(0.15, 0.2, 0.01, 0.06, 0.025))
}
\arguments{
\item{data}{the dataframe containing the data used by the function}

\item{ymax}{a vector defining the maximal values along the y axis (plot per plot)}
}
\description{
Build figure S1
}
\examples{
figureS1(data_all)

}
