% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{figureSX}
\alias{figureSX}
\title{Build figure SX}
\usage{
figureSX(best_fit)
}
\arguments{
\item{best_fit}{the fit of the model with the best predictive power}
}
\description{
Build figure SX
}
\examples{
#figureSX(best_fit)

}
