% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit the Generalised Linear Mixed Effects Models}
\usage{
fit(formulas, data, ncpus = 2)
}
\arguments{
\item{formulas}{the dataframe of formulas created with \code{prepare_formulas()}}

\item{data}{the dataframe containing the data used by the function}

\item{ncpus}{the number of CPU cores to use for parallel processing (default = 2)}
}
\value{
a dataframe with the formulas and all the statistics of goodness of fit
}
\description{
This functions fit the GLMMs predicting the return probability of the adult females.
}
\examples{
\dontrun{
  twoformulas <- prepare_formulas()[1:2,]
  fit(formulas = twoformulas, data = data_model)
}

}
