% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fit_comparison}
\alias{fit_comparison}
\title{The comparison of alternative fits}
\format{
a \code{tibble} object produced by \code{\link[=compare_fits]{compare_fits()}}
}
\description{
This object contains the output with the function \code{\link[=compare_fits]{compare_fits()}} as shown in \code{?mallaRd}.
}
\examples{
fit_comparison

}
\references{
this study
}
\seealso{
\code{\link[=compare_fits]{compare_fits()}}
}
\keyword{data}
