% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fitting_results}
\alias{fitting_results}
\title{The table of fitted the models}
\format{
a \code{tibble}
}
\description{
This tibble contains the models fitted with the function \code{\link[=fit]{fit()}} as shown in \code{?mallaRd}.
}
\examples{
fitting_results

}
\references{
this study
}
\seealso{
\code{\link[=fit]{fit()}}
}
\keyword{datasets}
