% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{howmany}
\alias{howmany}
\title{Compute the number of elements in a vector after discarding NAs}
\usage{
howmany(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the number of elements
}
\description{
Compute the number of elements in a vector after discarding NAs
}
\examples{
howmany(c(1, 3, NA))
howmany(c("1", "3", NA))

}
