% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Filtering and reformatting the data for statistical modelling}
\usage{
prepare_data(rawdata, filtering = TRUE)
}
\arguments{
\item{rawdata}{the data frame containing the raw data}

\item{filtering}{whether or not to apply filters on the data}
}
\value{
a dataframe with the data ready for modelling
}
\description{
This functions reformat the data so as to express predictors as differences between a capture
event and the previous capture. It can also perform a series of data filtering steps. Check the
body of the function to see all steps in details.
}
\examples{
#see ?mallaRd

}
