% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_formulas.R
\name{prepare_formulas}
\alias{prepare_formulas}
\title{Prepare the formulas for the models to be fitted}
\usage{
prepare_formulas()
}
\value{
a dataframe with the formulas of all models to be fitted
}
\description{
Prepare the formulas for the models to be fitted
}
\examples{
prepare_formulas()

}
