% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pretty}
\alias{pretty}
\title{Print numbers keeping a certain amount of significant digits}
\usage{
pretty(x, digits = 3)
}
\arguments{
\item{x}{a vector or a scalar}

\item{digits}{the number of digits to display}
}
\value{
a nicely displayed number
}
\description{
Print numbers keeping a certain amount of significant digits
}
