% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{summarize_fit}
\alias{summarize_fit}
\title{Compute stuff on the fit}
\usage{
summarize_fit(best_fit, method = "PQL/L", pretty = TRUE)
}
\arguments{
\item{best_fit}{the fit of the model with the best predictive power}

\item{method}{the fitting method ("PQL/L" or "PQL"), see \link[spaMM:method]{spaMM::method}}

\item{pretty}{whether or not to format the table for nice display}
}
\value{
a tibble with information on the fit
}
\description{
Compute stuff on the fit
}
\examples{
summarize_fit(best_fit)

}
