% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_biometrics}
\alias{table_biometrics}
\title{Create table with summary statistics about the biometrics of the mallards}
\usage{
table_biometrics(data)
}
\arguments{
\item{data}{the dataframe containing the data used by the function}
}
\value{
a dataframe
}
\description{
Create table with summary statistics about the biometrics of the mallards
}
\examples{
# example code
table_biometrics(data_all)

}
