% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table_predictors}
\alias{table_predictors}
\title{Create table with summary statistics about the predictors}
\usage{
table_predictors(data)
}
\arguments{
\item{data}{the dataframe containing the data used by the function}
}
\value{
a dataframe
}
\description{
Create table with summary statistics about the predictors
}
\examples{
# example code
table_predictors(data_model)

}
