% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{TjurD}
\alias{TjurD}
\title{Compute the Tjur's D statistics}
\usage{
TjurD(fit)
}
\arguments{
\item{fit}{a fitted model}
}
\value{
a dataframe with the Tjur's D and associated statistics
}
\description{
Tjur's D measures the accuracy of the fitted model at predicting the fitted data. D is defined by
the average difference between the probabilities of staying for breeding events that did actually
occur at the same location and for those that did not occur at the same locations. The predictive
power increases with increasing D.
}
\references{
Tjur, T. (2009). Coefficients of determination in logistic regression models - A new
proposal: The coefficient of discrimination. The American Statistician, 63(4), 366-372.
}
