% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_model}
\alias{data_model}
\title{The dataset used for fitting the models}
\format{
a \code{tibble}
}
\description{
This dataset contains the data used to fit all the models. It derives from \code{data_all} as shown in
\code{?mallaRd}.
}
\examples{
data_model

}
\references{
this study
}
\seealso{
\link{data_raw}, \link{data_all}
}
\keyword{datasets}
