% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distance2points}
\alias{distance2points}
\title{Compute the distance between two geographic locations}
\usage{
distance2points(lat1, long1, lat2, long2)
}
\arguments{
\item{lat1}{The latitude of point 1.}

\item{long1}{The longitude of point 1.}

\item{lat2}{The latitude of point 2.}

\item{long2}{The longitude of point 2.}
}
\value{
an object of class \code{units} providing the distance between two geographic locations.
}
\description{
Compute the distance between two geographic locations
}
\examples{
# Compute the distance between Paris and Berlin
distance2points(lat1 = 48.856667, long1 = 2.352222, lat2 = 52.52, long2 = 13.405)

}
