% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{figure1}
\alias{figure1}
\title{Build figure 1}
\usage{
figure1(data, ymax = c(0.15, 0.2, 0.01, 0.06, 0.025))
}
\arguments{
\item{data}{the dataframe containing the data used by the function}

\item{ymax}{a vector defining the maximal values along the y axis (plot per plot)}
}
\description{
Build figure 1
}
\examples{
figure1(data_all)

}
