% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{figure3}
\alias{figure3}
\title{Build figure 3}
\usage{
figure3(data_model, fit, ...)
}
\arguments{
\item{data_model}{the dataframe containing the data used for model fitting}

\item{fit}{a fitted model}

\item{...}{addition arguments for internal functions}
}
\description{
Build figure 3
}
\examples{
figure3(data_model, best_fit)

}
