% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallaRd-package.R
\docType{package}
\name{mallaRd-package}
\alias{mallaRd-package}
\alias{_PACKAGE}
\alias{mallaRd}
\title{mallaRd: Breeding output and breeding site fideliy of mallards in Berlin, Germany}
\description{
This help page reproduces, in the section \strong{Examples} the analyses and results from the paper
"Housing search in the concrete jungle – breeding site selection of urban mallards and
implications for conservation management strategies" by Engler et al.
}
\examples{
1 + 1

}
\references{
to come
}
\author{
\strong{Maintainer}: Alexandre Courtiol \email{alexandre.courtiol@gmail.com} (\href{https://orcid.org/0000-0003-0637-2959}{ORCID})

Authors:
\itemize{
  \item Marc Engler \email{mengler@nabu-berlin.de} (\href{https://orcid.org/0000-0002-3461-0953}{ORCID})
}

}
\keyword{package}
