% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Reformatting the data for statistical modelling}
\usage{
prepare_data(rawdata)
}
\arguments{
\item{rawdata}{the data frame containing the raw data}
}
\value{
a dataframe with the data ready for modeling
}
\description{
Reformatting the data for statistical modelling
}
\examples{
#see ?mallaRd

}
