% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Filtering and reformatting the data for statistical modelling}
\usage{
prepare_data(rawdata)
}
\arguments{
\item{rawdata}{the data frame containing the raw data}
}
\value{
a dataframe with the data ready for modeling
}
\description{
This functions perform a series of data filtering steps. It also express predictors as
differences between a capture event and the previous capture. Check the body of the function to
see all steps in details.
}
\examples{
#see ?mallaRd

}
