% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{subfigure4}
\alias{subfigure4}
\title{Build components of figure 4}
\usage{
subfigure4(data_model, fit, predictor, xlab = "")
}
\arguments{
\item{data_model}{the dataframe containing the data used for model fitting}

\item{fit}{a fitted model}

\item{predictor}{the name of the focal predictor}

\item{xlab}{the name for the x-axis}
}
\description{
Build components of figure 4
}
\examples{
subfigure4(data_model, best_fit, "relocation_distance_previous_z", "Relocation distance")

}
