% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{summarize_fit}
\alias{summarize_fit}
\title{Compute stuff on the fit}
\usage{
summarize_fit(best_fit)
}
\arguments{
\item{best_fit}{the fit of the model with the best predictive power}
}
\value{
a tibble with information on the fit
}
\description{
Compute stuff on the fit
}
\examples{
summarize_fit(best_fit)

}
